+++ src/xrEngine/CameraManager.h
@@ -100,7 +100,7 @@ public:
 #endif
 
 	void					Dump					();
-	u32						Count					()	{return m_EffectorsCam.size()+m_EffectorsCam_added_deffered.size();}
+	u32						Count					()	{return u32(m_EffectorsCam.size()+m_EffectorsCam_added_deffered.size());}
 	CEffectorCam*			AddCamEffector			(CEffectorCam*		ef);
 	CEffectorCam*			GetCamEffector			(ECamEffectorType	type);
 	void					RemoveCamEffector		(ECamEffectorType	type);
